refactor: change DIR
