+++ front-end/src/App.scss
@@ -1,4 +1,4 @@
-@import "components/variable/variable.scss";
+@import "components/variable/";
 
 html {
   font-size: 62.5%;
