--- front-end/src/components/Body/PostsPage/Card/style.scss
@@ -1,35 +1,0 @@
-@import "components/variable/variable.scss";
-
-.Card {
-  width: calc((100% - 8rem) / 4);
-  border: 1px solid rgba(0, 0, 0, 0.125);
-  border-radius: 4px;
-  margin: 1rem;
-  transition: all 200ms 0ms ease;
-  &:not(:first-child, :nth-child(4n + 1)) {
-    margin-left: 0;
-  }
-  :hover {
-    background: lighten(#c4ccf8, 10%);
-  }
-  figure {
-    margin: 0;
-  }
-  img {
-    max-width: 100%;
-  }
-  figcaption {
-    padding: 1rem;
-    :first-child {
-      font-size: 1.6rem;
-      font-weight: 500;
-      margin: 0 0 1rem 0;
-    }
-    :last-child {
-      font-weight: normal;
-      font-size: 1.2rem;
-      line-height: 150%;
-      margin-bottom: 0;
-    }
-  }
-}
