--- front-end/src/components/Body/PostsPage/style.scss
@@ -1,9 +1,0 @@
-@import "components/variable/variable.scss";
-
-.PostPage {
-  background: #fff;
-  display: flex;
-  width: 100%;
-  justify-content: space-around;
-  flex-wrap: wrap;
-}
