--- front-end/src/components/Body/style.scss
@@ -1,7 +1,0 @@
-@import "components/variable/variable.scss";
-
-.Body {
-  width: 100%;
-  position: relative;
-  padding: 1rem 0 1rem 1rem;
-}
