--- front-end/src/components/Header/User/style.scss
@@ -1,23 +1,0 @@
-@import "../../variable/variable.scss";
-
-.user {
-  display: flex;
-  align-items: center;
-  > * {
-    padding: 1rem;
-  }
-  img {
-    width: 4.5rem;
-    height: 4.5rem;
-    border-radius: 50%;
-  }
-  .user-profile {
-    :last-child {
-      color: $subTextColor;
-    }
-    > * {
-      margin: 0;
-      font-weight: 400;
-    }
-  }
-}
