--- front-end/src/components/Header/headerMenu/style.scss
@@ -1,10 +1,0 @@
-@import "components/variable/variable.scss";
-
-.icons {
-  display: flex;
-  align-items: center;
-  li {
-    list-style: none;
-    padding: 1rem;
-  }
-}
