--- front-end/src/components/Sidebar/style.scss
@@ -1,10 +1,0 @@
-@import "../variable/variable.scss";
-
-.side {
-  max-width: 20%;
-  background: $primaryBgColor;
-  .sid-sections {
-    padding: 0 2rem;
-    border-top: 1px solid #e9ecef;
-  }
-}
