+++ front-end/src/components/layoutComponents/Body/style.scss
@@ -1,0 +1,7 @@
+@import "components/variable/";
+
+.Body {
+  width: 100%;
+  position: relative;
+  padding: 1rem 0 1rem 1rem;
+}
