+++ front-end/src/components/layoutComponents/Footer/style.scss
@@ -1,0 +1,22 @@
+@import "components/variable";
+
+footer {
+  background: rgba(#fefefe, 0.9);
+  padding: 1rem 2rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  ul {
+    display: flex;
+    align-items: center;
+    list-style: none;
+    color: #000;
+    padding: 0;
+    span {
+      color: #0d6efd;
+    }
+    &.left :not(:first-child) {
+      margin-left: 2rem;
+    }
+  }
+}
