+++ front-end/src/components/layoutComponents/Sidebar/style.scss
@@ -1,0 +1,10 @@
+@import "../../variable/";
+
+.side {
+  max-width: 20%;
+  background: $primaryBgColor;
+  .sid-sections {
+    padding: 0 2rem;
+    border-top: 1px solid #e9ecef;
+  }
+}
