+++ front-end/src/components/pageComponents/PostsPage/style.scss
@@ -1,0 +1,9 @@
+@import "components/variable/";
+
+.PostPage {
+  background: #fff;
+  display: flex;
+  width: 100%;
+  justify-content: space-around;
+  flex-wrap: wrap;
+}
