+++ front-end/src/components/partsComponents/Card/index.jsx
