+++ front-end/src/components/partsComponents/Current/style.scss
@@ -1,0 +1,35 @@
+@import "components/variable/";
+
+.Current {
+  background: #fff;
+  padding: 1rem;
+  margin-bottom: 1rem;
+  border-radius: 8px;
+  p {
+    display: flex;
+    align-items: center;
+    &::before {
+      margin: 0 0 0 1rem;
+      content: "";
+      display: inline-block;
+      width: 2rem;
+      height: 2rem;
+      background: url("/asset/home.png") no-repeat center/cover;
+    }
+  }
+  span {
+    font-size: 1.4rem;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 24px;
+    letter-spacing: 0em;
+    text-align: left;
+    color: #6c757d;
+    margin: 0 1rem;
+    &:first-child {
+      color: #0d6efd;
+      text-decoration: underline;
+      text-decoration-color: #0d6efd;
+    }
+  }
+}
