+++ front-end/src/components/partsComponents/Pagination/index.jsx
