+++ front-end/src/components/partsComponents/SideSection/SideSectionItem/index.jsx
