+++ front-end/src/components/partsComponents/SideSection/index.jsx
