+++ front-end/src/components/partsComponents/User/style.scss
@@ -1,0 +1,23 @@
+@import "../../variable/";
+
+.user {
+  display: flex;
+  align-items: center;
+  > * {
+    padding: 1rem;
+  }
+  img {
+    width: 4.5rem;
+    height: 4.5rem;
+    border-radius: 50%;
+  }
+  .user-profile {
+    :last-child {
+      color: $subTextColor;
+    }
+    > * {
+      margin: 0;
+      font-weight: 400;
+    }
+  }
+}
