+++ front-end/src/components/partsComponents/UserRow/style.scss
@@ -1,0 +1,32 @@
+@import "components/variable/";
+
+td {
+  &:first-child {
+    text-align: center;
+    width: 5%;
+  }
+  &:nth-child(2) {
+    width: 5%;
+  }
+  &:nth-child(3) {
+    width: calc(80% / 3);
+  }
+  &:nth-child(4) {
+    width: calc(80% / 3);
+  }
+  &:nth-child(5) {
+    width: calc(80% / 3);
+  }
+  &:last-child {
+    color: #0d6efd;
+    padding: 0.5rem;
+    text-align: center;
+    width: 10%;
+    a + a {
+      margin-left: 1rem;
+    }
+  }
+  &:not(:first-child, :last-child) {
+    padding: 0.5rem 1rem;
+  }
+}
