+++ front-end/src/components/partsComponents/headerMenu/style.scss
@@ -1,0 +1,10 @@
+@import "components/variable/";
+
+.icons {
+  display: flex;
+  align-items: center;
+  li {
+    list-style: none;
+    padding: 1rem;
+  }
+}
