+++ front-end/src/components/variable/index.scss
