--- front-end/src/components/variable/variable.js
@@ -1,2 +1,0 @@
-const BODY_DIR = "components/Body";
-export default BODY_DIR;
